split CDK stack AOT and non-AOT. refactor. Initial GH actions pipelines
