+++ libraries/tests/e2e/functions/core/metrics/Function/test/Function.Tests/FunctionTest.cs
@@ -57,11 +57,11 @@ public class FunctionTest
             Assert.Equal("HELLO WORLD", parsedPayload.Body);
 
             // Assert Output log from Lambda execution
-            AssertOutputLog(functionName, response);
+            AssertOutputLog(response);
         }
     }
 
-    private void AssertOutputLog(string functionName, InvokeResponse response)
+    private void AssertOutputLog(InvokeResponse response)
     {
         // Extract and parse log
         var logResult = System.Text.Encoding.UTF8.GetString(Convert.FromBase64String(response.LogResult));
@@ -85,7 +85,7 @@ public class FunctionTest
         JsonElement root = doc.RootElement;
 
         Assert.True(root.TryGetProperty("_aws", out JsonElement awsElement));
-        
+
         Assert.True(awsElement.TryGetProperty("CloudWatchMetrics", out JsonElement cloudWatchMetricsElement));
         Assert.True(cloudWatchMetricsElement[0].TryGetProperty("Namespace", out JsonElement namespaceElement));
         Assert.Equal("Test", namespaceElement.GetString());
