+++ libraries/tests/e2e/infra-aot/CoreAotStack.cs
@@ -4,7 +4,7 @@ using Amazon.CDK.AWS.Lambda;
 using Constructs;
 using Architecture = Amazon.CDK.AWS.Lambda.Architecture;
 
-namespace E2E;
+namespace InfraAot;
 
 public class CoreAotStack : Stack
 {
@@ -22,16 +22,16 @@ public class CoreAotStack : Stack
 
         if (RuntimeInformation.OSArchitecture == System.Runtime.InteropServices.Architecture.Arm64)
         {
-            CreateFunctionConstruct(this, $"{utility}_ARM_aot_net8", Runtime.DOTNET_8, Architecture.ARM_64, $"E2ETestLambda_ARM_AOT_NET8_{utility}", baseAotPath, distAotPath, true);
+            CreateFunctionConstruct(this, $"{utility}_ARM_aot_net8", Runtime.DOTNET_8, Architecture.ARM_64, $"E2ETestLambda_ARM_AOT_NET8_{utility}", baseAotPath, distAotPath);
         }
 
         if (RuntimeInformation.OSArchitecture == System.Runtime.InteropServices.Architecture.X64)
         {
-            CreateFunctionConstruct(this, $"{utility}_X64_aot_net8", Runtime.DOTNET_8, Architecture.X86_64, $"E2ETestLambda_X64_AOT_NET8_{utility}", baseAotPath, distAotPath, true);
+            CreateFunctionConstruct(this, $"{utility}_X64_aot_net8", Runtime.DOTNET_8, Architecture.X86_64, $"E2ETestLambda_X64_AOT_NET8_{utility}", baseAotPath, distAotPath);
         }
     }
 
-    private void CreateFunctionConstruct(Construct scope, string id, Runtime runtime, Architecture architecture, string name, string sourcePath, string distPath, bool isAot)
+    private void CreateFunctionConstruct(Construct scope, string id, Runtime runtime, Architecture architecture, string name, string sourcePath, string distPath)
     {
         new FunctionConstruct(scope, id, new FunctionConstructProps
         {
@@ -40,8 +40,7 @@ public class CoreAotStack : Stack
             Name = name,
             Handler = "AOT-Function",
             SourcePath = sourcePath,
-            DistPath = distPath,
-            IsAot = isAot
+            DistPath = distPath
         });
     }
 }
