+++ libraries/tests/e2e/infra-aot/InfraAot.csproj
