+++ libraries/tests/e2e/infra-aot/InfraAotStack.cs
@@ -1,0 +1,13 @@
+using Amazon.CDK;
+using Constructs;
+
+namespace InfraAot
+{
+    public class InfraAotStack : Stack
+    {
+        internal InfraAotStack(Construct scope, string id, IStackProps props = null) : base(scope, id, props)
+        {
+            // The code that defines your stack goes here
+        }
+    }
+}
