+++ libraries/tests/e2e/infra-aot/Program.cs
@@ -1,15 +1,16 @@
 ﻿using Amazon.CDK;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 
-namespace Infra
+namespace InfraAot
 {
     sealed class Program
     {
         public static void Main(string[] args)
         {
             var app = new App();
-
-            new CoreStack(app, "CoreStack", new StackProps { });
-
+            new CoreAotStack(app, "CoreAotStack", new StackProps { });
             app.Synth();
         }
     }
